Circleci project setup

+++ src/test/java/jl/project/screenreaderstests/SpeechToTextConverterIBM.java
@@ -1,25 +1,37 @@
-// https://www.ibm.com/cloud/watson-speech-to-textError stream curl: (1) Protocol ""https" not supported or disabled in libcurl
+// https://www.ibm.com/cloud/watson-speech-to-text
 // https://cloud.ibm.com/services/speech-to-text/crn%3Av1%3Abluemix%3Apublic%3Aspeech-to-text%3Aeu-gb%3Aa%2F2e4a017664644b6c865cf4e532dd9056%3A08dcb5ed-d57f-410f-98ef-11ae645b5a04%3A%3A?paneId=manage&new=true
 // https://cloud.ibm.com/docs/speech-to-text?topic=speech-to-text-gettingStarted#getting-started-tutorial
 
 package jl.project.screenreaderstests;
 
+import java.awt.AWTException;
 import java.awt.Robot;
+import java.io.File;
+import java.io.IOException;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import jl.project.StringExternalization;
+import jl.project.exceptions.UnrecognizedOSException;
 
 public class SpeechToTextConverterIBM 
 {
 	Logger logger = LoggerFactory.getLogger(SpeechToTextConverterIBM.class);
 	Robot robot;
 	ProcessBuilder processBuilder = new ProcessBuilder();	
+	String root_folder = "src/test/java/jl/project/screenreaderstests/";
+	File outputLog = new File("logs/output.txt");
+	File errorLog = new File("logs/error.txt");
 	String apikeyParam="\"apikey:"+System.getenv("IBM_SpeechToText")+"\"";	
 	
 	
 	
-	public String convertAudioToText(String pathToAudioFile) throws Exception
+	public String convertAudioToText(String audioFileName) throws UnrecognizedOSException
 	{
+		System.setProperty("pathToAudioFile", audioFileName);
+		logger.debug(String.format("Value for the system property 'pathToAudioFile': %s", System.getProperty("pathToAudioFile")));
+		 
 		//Needed to fix an new line issue in the environment variable value in Windows.
 		apikeyParam = apikeyParam.replace("\n", "");
 		
@@ -28,12 +40,16 @@ public class SpeechToTextConverterIBM
 		String osName = System.getProperty("os.name");
 		Process process = null;
 		
-		robot = new Robot();
+		try {
+			robot = new Robot();
+		} catch (AWTException e) {
+			logger.debug(StringExternalization.EXCEPTION_AWT);
+		}
 		
 		if (logger.isDebugEnabled())logger.debug(String.format("OS: %s",osName));		
 		if(osName.contains("Windows"))
 		{
-			processBuilder.command("src/test/java/jl/project/ScreenReadersTests/scripts/script_IBM_STT_Windows.bat");
+			processBuilder.command(String.format("src/test/java/jl/project/ScreenReadersTests/scripts/script_IBM_STT_Windows.bat %s",audioFileName));
 		}
 		else if (osName.contains("Mac"))
 		{
@@ -44,27 +60,37 @@ public class SpeechToTextConverterIBM
 			processBuilder.command("src/test/java/jl/project/ScreenReadersTests/scripts/script_IBM_STT_Ubuntu.sh");
 
 		}
-		else {throw new Exception(String.format("Unrecognized OS : %s",osName));}
+		else {throw new UnrecognizedOSException(String.format("Unrecognized OS : %s",osName));}
 		
-		process = processBuilder.start();
-		logger.debug("Process started");
+		try 
+		{
+			processBuilder.redirectInput(outputLog);
+			processBuilder.redirectError(errorLog);
+			process = processBuilder.start();
 		
-		logger.debug("Waiting for results.");
-		robot.delay(15000);
-		
-		//logger.debug(String.format("Error stream %s",new String(process.getErrorStream().readAllBytes())));		
-		//Stream that gives the output 		
-		//logger.debug(String.format("Input stream %s",new String(process.getInputStream().readAllBytes())));
-		
-		
-		// readAllBytes() since Java 1.9
-		// https://docs.oracle.com/javase/9/docs/api/java/io/InputStream.html#readAllBytes--
-		informationReturned = new String(process.getInputStream().readAllBytes());
-		
+			logger.debug("Request sent.");
+			
+			logger.debug("Waiting for results.");
+			robot.delay(15000);
+			
+			// readAllBytes() since Java 1.9
+			// https://docs.oracle.com/javase/9/docs/api/java/io/InputStream.html#readAllBytes--
+			informationReturned = new String(process.getInputStream().readAllBytes());
+			
+			logger.debug(String.format("Value for the system property 'pathToAudioFile': %s", System.getProperty("pathToAudioFile")));
+		} 
+		catch (IOException e) 
+		{
+			logger.debug(StringExternalization.EXCEPTION_IO);
+		}
 		// supportsNormalTermination() since Java 1.9
 		// https://docs.oracle.com/javase/9/docs/api/java/lang/Process.html#supportsNormalTermination--
-		if (process.supportsNormalTermination()) process.destroy(); else process.destroyForcibly();
-		if (logger.isDebugEnabled()) logger.debug(String.format("Process is alive : %b", process.isAlive()));				
+		if (process != null)
+		{
+			if ( process.supportsNormalTermination()) process.destroy(); else process.destroyForcibly();
+			if (logger.isDebugEnabled()) logger.debug(String.format("Process is alive : %b", process.isAlive()));		
+		}
+						
 		
 		return informationReturned;
 	}
